Create Binary Search.
Binary search algorithm searches the target value within a sorted array.

To perform a binary search array must be sorted, it should either be in ascending or descending order.
Step 1: First divide the list of elements in half.
Step 2: In the second step we compare the target value with the middle element of the array. If it matches the middle element then search ends here and doesn’t proceed further.
Step 3: Else if the element less than the middle element then we begin the search in lower half of the array.
Step 4: Else if the element is greater than the middle element then we begin the search in greater half of the array.
Step 5: We will repeat the steps 3 ,4 and 5 until our target element is found.